Move Follow list for Follow Page to a separate component
