+++ src/screens/Profiles/components/FollowList/index.js
@@ -1,0 +1,1 @@
+export * from './component';
